+++ etc/blocks/pool-explorer/block/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @milaboratories/milaboratories.pool-explorer
 
+## 1.0.314
+
+### Patch Changes
+
+- @milaboratories/milaboratories.pool-explorer.workflow@1.1.76
+- @milaboratories/milaboratories.pool-explorer.ui@1.1.214
+
 ## 1.0.313
 
 ### Patch Changes